Move mutations to file
